refactor(admob): use Jetpack's Navigation Component
