+++ admob/app/build.gradle
@@ -28,8 +28,8 @@ android {
     }
 
     buildFeatures {
-  viewBinding = true
-}
+        viewBinding = true
+    }
 }
 
 dependencies {
@@ -38,6 +38,8 @@ dependencies {
     implementation 'androidx.appcompat:appcompat:1.2.0'
     implementation 'com.google.android.material:material:1.2.1'
     implementation 'androidx.browser:browser:1.0.0'
+    implementation 'androidx.navigation:navigation-fragment-ktx:2.3.2'
+    implementation 'androidx.navigation:navigation-ui-ktx:2.3.2'
 
     // [START gradle_play_config]
     implementation 'com.google.android.gms:play-services-ads:19.6.0'
