--- admob/app/src/main/java/com/google/samples/quickstart/admobexample/java/SecondActivity.java
@@ -1,32 +1,0 @@
-/**
- * Copyright Google Inc. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.google.samples.quickstart.admobexample.java;
-
-import android.os.Bundle;
-import androidx.appcompat.app.AppCompatActivity;
-
-import com.google.samples.quickstart.admobexample.R;
-
-
-public class SecondActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.second_activity);
-    }
-
-}
