+++ admob/app/src/main/java/com/google/samples/quickstart/admobexample/java/SecondFragment.java
@@ -1,0 +1,17 @@
+package com.google.samples.quickstart.admobexample.java;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import androidx.fragment.app.Fragment;
+import com.google.samples.quickstart.admobexample.R;
+
+class SecondFragment extends Fragment {
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+        return inflater.inflate(R.layout.fragment_second, container, false);
+    }
+
+}
