--- admob/app/src/main/java/com/google/samples/quickstart/admobexample/kotlin/SecondActivity.kt
@@ -1,13 +1,0 @@
-package com.google.samples.quickstart.admobexample.kotlin
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import com.google.samples.quickstart.admobexample.R
-
-class SecondActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.second_activity)
-    }
-}
