+++ admob/app/src/main/java/com/google/samples/quickstart/admobexample/kotlin/SecondFragment.kt
@@ -1,0 +1,16 @@
+package com.google.samples.quickstart.admobexample.kotlin
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.google.samples.quickstart.admobexample.R
+
+class SecondFragment : Fragment() {
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+        return inflater.inflate(R.layout.fragment_second, container, false)
+    }
+
+}
