+++ admob/app/src/main/res/layout/fragment_second.xml
@@ -1,9 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.google.samples.quickstart.admobexample.java.SecondActivity"
     tools:viewBindingIgnore="true">
 
     <ImageView
@@ -21,7 +21,7 @@
         android:layout_height="wrap_content"
         android:layout_below="@+id/icon"
         android:layout_gravity="center_horizontal"
-        android:text="@string/second_activity_content"
+        android:text="@string/second_fragment_content"
         android:textAlignment="center"
         app:layout_constraintEnd_toEndOf="@+id/icon"
         app:layout_constraintStart_toStartOf="@+id/icon"
