+++ admob/app/src/main/res/navigation/nav_graph.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/nav_graph"
+    app:startDestination="@id/FirstFragment">
+
+    <fragment
+        android:id="@+id/FirstFragment"
+        android:name="com.google.samples.quickstart.admobexample.kotlin.FirstFragment"
+        android:label="@string/first_fragment_title"
+        tools:layout="@layout/fragment_first">
+
+        <action
+            android:id="@+id/action_FirstFragment_to_SecondFragment"
+            app:destination="@id/SecondFragment" />
+    </fragment>
+    <fragment
+        android:id="@+id/SecondFragment"
+        android:name="com.google.samples.quickstart.admobexample.kotlin.SecondFragment"
+        android:label="@string/second_fragment_title"
+        tools:layout="@layout/fragment_second">
+    </fragment>
+</navigation>
