+++ admob/app/src/main/res/values/strings.xml
@@ -2,8 +2,9 @@
     <string name="app_name">AdMob Quickstart</string>
     <string name="hello_world">Hello world!</string>
     <string name="interstitial_button_text">Show Interstitial Ad</string>
-    <string name="second_activity_title">SecondActivity</string>
-    <string name="second_activity_content">This is the second activity.</string>
+    <string name="first_fragment_title">FirstFragment</string>
+    <string name="second_fragment_title">SecondFragment</string>
+    <string name="second_fragment_content">This is the second fragment.</string>
 
     <!--
         TODO: Replace this with your own Admob App ID! This value is only a sample.
